chore: move cache into client/ internal/cache
以此缓存仅在 client 内部使用故
