+++ client/internal/cache/base.go
