+++ client/internal/cache/operation.go
