--- message/doc.go
@@ -1,3 +1,0 @@
-package message
-
-// from https://github.com/Mrs4s/MiraiGo/tree/master/message
